Atualização
...

+++ src/com/mycompany/usuarios/UsuarioRN.java
@@ -23,7 +23,7 @@ public class UsuarioRN {
 
 	
 	public void salvar(UsuarioEntity usuario) {
-		Integer codigo = usuario.getCodigo();
+		Long codigo = usuario.getCodigo();
 		if (codigo == null || codigo == 0) {
 	//		usuario.getPermissao().add("ROLE_USUARIO"); 
 			this.usuarioDAO.salvar(usuario);
